fix bug

+++ app/Http/Controllers/ArtistController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use App\Artist;
 use App\Image;
+use App\Video;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\DB;
@@ -226,11 +227,45 @@ class ArtistController extends Controller
      */
     public function destroy($artist)
     {
+        $video = Video::where('artist_id',$artist)->get();
+        $art = Artist::where('id',$artist)->first();
+
+        if ($art['image'] != null) {
+            $image_path = public_path() . '/storage/' . $art['image'];
+            unlink($image_path);
+        }
+        if ($art['video'] != null) {
+            $image_path = public_path() . '/storage/' . $art['video'];
+            unlink($image_path);
+        }
         Artist::destroy($artist);
 
+        if (!empty($video)) {
+            foreach ($video as $value) {
+                DB::table('advertise')
+                    ->where('video_id', $value->id)
+                    ->delete();
+                if ($value->image != null) {
+                    $image_path = public_path() . '/storage/' . $value->image;
+                    unlink($image_path);
+                }
+                if ($value->video != null) {
+                    $image_path = public_path() . '/storage/' . $value->video;
+                    unlink($image_path);
+                }
+            }
+        }
         DB::table('video')->where('artist_id',$artist)->delete();
 
-        DB::table('image')->where('artist_id',$artist)->delete();
+        $image = Image::where('artist_id', $artist)->get();
+        if (!empty($image)) {
+            foreach ($image as $item) {
+                $image_path = public_path() . '/storage/' . $item->image;
+                unlink($image_path);
+            }
+            DB::table('image')->where('artist_id',$artist)->delete();
+        }
+
         return redirect('artist');
     }
 }
